+++ packages/backend/src/services/ProjectService/ProjectService.ts
@@ -111,6 +111,7 @@ import {
     getMetrics,
     getTimezoneLabel,
     hasIntersection,
+    isAndFilterGroup,
     isCustomSqlDimension,
     isDateItem,
     isDimension,
@@ -118,6 +119,7 @@ import {
     isFilterRule,
     isFilterableDimension,
     isNotNull,
+    isOrFilterGroup,
     isUserWithOrg,
     maybeReplaceFieldsInChartVersion,
     replaceDimensionInExplore,
@@ -1614,6 +1616,7 @@ export class ProjectService extends BaseService {
             ...addDashboardFiltersToMetricQuery(
                 savedChart.metricQuery,
                 appliedDashboardFilters,
+                explore,
             ),
             sorts:
                 dashboardSorts && dashboardSorts.length > 0