+++ packages/common/src/compiler/filtersCompiler.ts
@@ -144,6 +144,14 @@ export const renderNumberFilterSql = (
             return `(${dimensionSql}) < (${filter.values?.[0] || 0})`;
         case FilterOperator.LESS_THAN_OR_EQUAL:
             return `(${dimensionSql}) <= (${filter.values?.[0] || 0})`;
+        case FilterOperator.IN_BETWEEN:
+            return `(${dimensionSql}) >= (${
+                filter.values?.[0] || 0
+            }) AND (${dimensionSql}) <= (${filter.values?.[1] || 0})`;
+        case FilterOperator.NOT_IN_BETWEEN:
+            return `(${dimensionSql}) < (${
+                filter.values?.[0] || 0
+            }) OR (${dimensionSql}) > (${filter.values?.[1] || 0})`;
         default:
             return raiseInvalidFilterError('number', filter);
     }