+++ packages/frontend/src/ee/providers/Embed/context.ts
@@ -5,6 +5,7 @@ const EmbedProviderContext = createContext<EmbedContext>({
     embedToken: undefined,
     filters: undefined,
     projectUuid: undefined,
+    languageMap: undefined,
     t: (_input: string) => undefined,
 });